Removing stickyScrollController.ts duplicate file
